update tox tests
