+++ .pre-commit-config.yaml
@@ -1,0 +1,26 @@
+repos:
+- repo: https://github.com/ambv/black
+  rev: 21.5b1
+  hooks:
+  - id: black
+    name: black - Format code (-l 120)
+    language_version: python3.7
+    args: [-l 120]
+- repo: https://gitlab.com/pycqa/flake8
+  rev: 3.9.2
+  hooks:
+    - id: flake8
+# - repo: local
+#   hooks:
+#   - id: mypy
+#     name: mypy - Check static types match
+#     entry: pipenv run mypy src --ignore-missing-imports
+#     pass_filenames: false
+#     language: system
+#     types: [python]
+#   - id: pytest
+#     name: pytest - Check unit tests --ignore=socrates/vendors
+#     entry: pipenv run pytest --doctest-modules
+#     pass_filenames: false
+#     language: system
+#     types: [python]
