+++ Makefile
@@ -11,7 +11,7 @@ PYTHON_INTERPRETER = python
 ifeq (,$(shell which conda))
 HAS_CONDA=False
 else
-HAS_CONDA=True
+HAS_CONDA=False
 endif
 
 #################################################################################
@@ -63,7 +63,7 @@ conda_stop:
 ## Set up python interpreter environment
 create_environment:
 ifeq (True,$(HAS_CONDA))
-		@echo ">>> Detected conda, creating conda environment."
+	@echo ">>> Detected conda, creating conda environment."
 ifeq (3,$(findstring 3,$(PYTHON_INTERPRETER)))
 	conda create --name $(PROJECT_NAME) python=3.7
 else
