+++ requirements.txt
@@ -14,9 +14,9 @@ joblib==1.1.0
 langcodes==3.3.0
 MarkupSafe==2.1.1
 murmurhash==1.0.7
-numpy==1.23.0
+numpy==1.20.0
 packaging==21.3
-pandas==1.4.3
+pandas==1.3.3
 pathy==0.6.2
 preshed==3.0.6
 pyahocorasick==1.4.4
@@ -26,8 +26,8 @@ pyparsing==3.0.9
 python-dateutil==2.8.2
 pytz==2022.1
 requests==2.28.1
-scikit-learn==1.1.1
-scipy==1.8.1
+scikit-learn==1.0.1
+scipy==1.7.3
 six==1.16.0
 smart-open==5.2.1
 spacy==3.3.1
@@ -39,7 +39,7 @@ thinc==8.0.17
 threadpoolctl==3.1.0
 tqdm==4.64.0
 typer==0.4.1
-typing_extensions==4.2.0
+typing_extensions==4.1.0
 urllib3==1.26.9
 wasabi==0.9.1
 xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.3.0/xx_ent_wiki_sm-3.3.0-py3-none-any.whl
