+++ src/transform/nlp_operator.py
@@ -1,3 +1,4 @@
+import re
 from typing import Any, Dict, List
 
 import emot
@@ -486,14 +487,108 @@ class NlpReplaceEmoticons(BaseEstimator):
         return x
 
 
+class NlpDeDuplicatesSpace(BaseEstimator):
+
+    def __init__(self, text_column: str, new_column: str = None) -> None:
+        self.text_column = text_column
+        if new_column is None:
+            self.new_column = text_column
+        else:
+            self.new_column = new_column
+
+    def fit(self, x: Any, y: Any = None) -> __qualname__:
+        return self
+
+    @staticmethod
+    def remove_multiple_spaces(text: str):
+        return re.sub(" {2,}", " ", text)
+
+    def transform(self, x: Any) -> pd.DataFrame:
+        x[self.new_column] = x[self.text_column].map(self.remove_multiple_spaces)
+
+        return x
+
+
 class NlpRemoveCharacterRepetition(BaseEstimator):
-    pass
+
+    def __init__(self, text_column: str, new_column: str = None) -> None:
+        self.text_column = text_column
+        if new_column is None:
+            self.new_column = text_column
+        else:
+            self.new_column = new_column
+
+    def fit(self, x: Any, y: Any = None) -> __qualname__:
+        return self
+
+    @staticmethod
+    def remove_multiple_spaces(text: str):
+        return re.sub(" {2,}", " ", text)
+
+    def transform(self, x: Any) -> pd.DataFrame:
+        x[self.new_column] = x[self.text_column].map(self.remove_multiple_spaces)
+
+        return x
 
 
-class NlpRemoveWordRepetition(BaseEstimator):
-    pass
+
+class NlpReplaceWordRepetition(BaseEstimator):
+
+    def __init__(self, text_column: str, new_column: str = None) -> None:
+        self.text_column = text_column
+        if new_column is None:
+            self.new_column = text_column
+        else:
+            self.new_column = new_column
+
+    def fit(self, x: Any, y: Any = None) -> __qualname__:
+        return self
+
+    @staticmethod
+    def _replace_group(match):
+        char, _ = match.groups()
+
+        return char
+
+    def replace_words_rep(self, text: str) -> str:
+        word_rep = re.compile(r"(\b\w+\W+)(\1{2,})")
+        
+        return word_rep.sub(self._replace_group, text)
+
+    def transform(self, x: Any) -> pd.DataFrame:
+        x[self.new_column] = x[self.text_column].map(self.replace_words_rep)
+
+        return x
 
 
+
+class NlpRemoveCharRepetition(BaseEstimator):
+
+    def __init__(self, text_column: str, new_column: str = None) -> None:
+        self.text_column = text_column
+        if new_column is None:
+            self.new_column = text_column
+        else:
+            self.new_column = new_column
+
+    def fit(self, x: Any, y: Any = None) -> __qualname__:
+        return self
+
+    @staticmethod
+    def _replace_group(match):
+        char, _ = match.groups()
+
+        return char
+        
+    def replace_char_rep(self, text: str) -> str:
+        char_rep = re.compile(r"(\S)(\1{2,})")
+        
+        return char_rep.sub(self._replace_group, text)
+
+    def transform(self, x: Any) -> pd.DataFrame:
+        x[self.new_column] = x[self.text_column].map(self.replace_char_rep)
+
+        return x
     # @staticmethod
     # def _replace_group(match):
     #     char, repetition = match.groups()
