+++ tox.ini
@@ -1,3 +1,11 @@
 [flake8]
-max-line-length = 79
+max-line-length = 120
 max-complexity = 10
+
+[tox]
+envlist = py37
+
+[testenv]
+deps = -r{toxinidir}/requirements.txt
+       ; -r{toxinidir}/requirements.testing.txt
+commands = pytest {posargs:tests}
